fix(Format) fix rebase artifacts and new places where uint32_t prepended size appears

+++ nes-nautilus/src/Nautilus/Arena.cpp
@@ -82,7 +82,7 @@ nautilus::val<int8_t*> ArenaRef::allocateMemory(const nautilus::val<size_t>& siz
     return currentArenaPtr;
 }
 
-VariableSizedData ArenaRef::allocateVariableSizedData(const nautilus::val<uint32_t>& sizeInBytes)
+VariableSizedData ArenaRef::allocateVariableSizedData(const nautilus::val<uint64_t>& sizeInBytes)
 {
     const auto basePtr = allocateMemory(sizeInBytes);
     return VariableSizedData(basePtr, sizeInBytes);
